move `backend-fetch` to `services/electron-ipc`
